+++ R/ignore.R
@@ -27,5 +27,6 @@ use_build_ignore <- function(files, escape = TRUE, base_path = ".") {
 
 escape_path <- function(x) {
   x <- gsub("\\.", "\\\\.", x)
+  x <- gsub("/$", "", x)
   paste0("^", x, "$")
 }