+++ tests/testthat/test-use_build_ignore.R
@@ -3,3 +3,7 @@ context("use_build_ignore")
 test_that(". escaped around surround by anchors", {
   expect_equal(escape_path("."), "^\\.$")
 })
+
+test_that("strip trailing /", {
+  expect_equal(escape_path("./"), "^\\.$")
+})