+++ NEWS.md
@@ -1,3 +1,5 @@
 # usethis 0.0.0.9000
 
 * Removed old `add_build_ignore()`
+
+* `use_build_ignore()` now strips trailing `/`